Sight Update
Suspicious-level added to the properties of the criminals and the passengers.

The security consider this variable when updating their memories of a given target.

Printout added to the start of the "look" function, and removed from the end. So that you know which agent is calling it, before the updated list is printed.